Verification integrer au script des actions + navigation amelioree

+++ administration/action-supprimer-jeu.php
@@ -1,18 +1,17 @@
 <?php
 	//print_r($_POST);
-	
-	$idJeu = $_POST["id"];
-	
-	if(!empty($_POST["confirmation-oui"])){
-		$SQL_EFFACER_JEU = "DELETE FROM jeu WHERE idJeu = " .$idJeu;
-	
-		include_once "base-de-donnees.php";
+	if(!empty($_POST["action-supprimer-jeu"])){
+		$idJeu = $_POST["id"];
 		
-		//echo $SQL_EFFACER_JEU;
+		if(!empty($_POST["confirmation-oui"])){
+			$SQL_EFFACER_JEU = "DELETE FROM jeu WHERE idJeu = " .$idJeu;
 		
-		$requeteEffacerJeu = $pdo->prepare($SQL_EFFACER_JEU);
-		$requeteEffacerJeu->execute();
+			include_once "base-de-donnees.php";
+			
+			//echo $SQL_EFFACER_JEU;
+			
+			$requeteEffacerJeu = $pdo->prepare($SQL_EFFACER_JEU);
+			$requeteEffacerJeu->execute();
+		}
 	}
-	
-	
 ?>
